Add password protected for hidden content
In order to mark a content as hidden add [p] prefix to it in plex.
Content can be entire section, playlist or individual clip.
in case a clip is mark as hidden and is placed in a playlist , the
playlist is considered as hidden.

How to set a password:
on the main menu hit the blue button:
Enter a 5 color button sequence (excluding blue) and confirm

How to see hidden content:
Type the blue button again in the main menu and enter the password

Panic button is red and can be clicked on all screens of the app causing
the app to close and set next start to unauthorized session (no hidden
content will be seen).